+++ lib/features/steps/header_steps.rb
@@ -6,20 +6,24 @@
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input header_name [String] The header to test
-Then('the {word} {string} header is not null') do |request_type, header_name|
+Then('the {request_type} {string} header is present') do |request_type, header_name|
   Maze.check.not_nil(Maze::Server.list_for(request_type).current[:request][header_name],
                      "The #{request_type} '#{header_name}' header should not be null")
+
+  request = Maze::Server.list_for(request_type).current[:request]
+  Maze.check.true(request.header.key?(header_name.downcase),
+                  "The #{request_type} '#{header_name}' header should be present")
 end
 
 # Tests that a request header is null
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input header_name [String] The header to test
-Then('the {word} {string} header is null') do |request_type, header_name|
+Then('the {request_type} {string} header is not present') do |request_type, header_name|
   request = Maze::Server.list_for(request_type).current[:request]
 
-  Maze.check.nil(request[header_name],
-                 "The #{request_type} '#{header_name}' header should be null")
+  Maze.check.false(request.header.key?(header_name.downcase),
+                   "The #{request_type} '#{header_name}' header should not be present")
 end
 
 # Tests that request header equals a string
@@ -27,7 +31,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input header_name [String] The header to test
 # @step_input header_value [String] The string it should match
-Then('the {word} {string} header equals {string}') do |request_type, header_name, header_value|
+Then('the {request_type} {string} header equals {string}') do |request_type, header_name, header_value|
   Maze.check.not_nil(Maze::Server.list_for(request_type).current[:request][header_name],
                      "The #{request_type} '#{header_name}' header wasn't present in the request")
   Maze.check.equal(header_value, Maze::Server.list_for(request_type).current[:request][header_name])
@@ -38,7 +42,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input header_name [String] The header to test
 # @step_input regex_string [String] The regex to match with
-Then('the {word} {string} header matches the regex {string}') do |request_type, header_name, regex_string|
+Then('the {request_type} {string} header matches the regex {string}') do |request_type, header_name, regex_string|
   regex = Regexp.new(regex_string)
   value = Maze::Server.list_for(request_type).current[:request][header_name]
   Maze.check.match(regex, value)
@@ -49,7 +53,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input header_name [String] The header to test
 # @step_input header_values [DataTable] A parsed data table
-Then('the {word} {string} header equals one of:') do |request_type, header_name, header_values|
+Then('the {request_type} {string} header equals one of:') do |request_type, header_name, header_values|
   Maze.check.include(header_values.raw.flatten, Maze::Server.list_for(request_type).current[:request][header_name])
 end
 
@@ -57,7 +61,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input header_name [String] The header to test
-Then('the {word} {string} header is a timestamp') do |request_type, header_name|
+Then('the {request_type} {string} header is a timestamp') do |request_type, header_name|
   header = Maze::Server.list_for(request_type).current[:request][header_name]
   Maze.check.match(TIMESTAMP_REGEX, header)
 end
@@ -65,7 +69,7 @@ end
 # Checks that the Bugsnag-Integrity header is a SHA1 or simple digest
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
-When('the {word} Bugsnag-Integrity header is valid') do |request_type|
+When('the {request_type} Bugsnag-Integrity header is valid') do |request_type|
   Maze.check.true(Maze::Helper.valid_bugsnag_integrity_header(Maze::Server.list_for(request_type).current),
                   'Invalid Bugsnag-Integrity header detected')
 end