+++ lib/features/steps/payload_steps.rb
@@ -6,7 +6,7 @@
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input fixture_path [String] Path to a JSON fixture
-Then('the {word} payload body does not match the JSON fixture in {string}') do |request_type, fixture_path|
+Then('the {request_type} payload body does not match the JSON fixture in {string}') do |request_type, fixture_path|
   payload_value = Maze::Server.list_for(request_type).current[:body]
   expected_value = JSON.parse(open(fixture_path, &:read))
   result = Maze::Compare.value(expected_value, payload_value)
@@ -17,7 +17,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input fixture_path [String] Path to a JSON fixture
-Then('the {word} payload body matches the JSON fixture in {string}') do |request_type, fixture_path|
+Then('the {request_type} payload body matches the JSON fixture in {string}') do |request_type, fixture_path|
   payload_value = Maze::Server.list_for(request_type).current[:body]
   expected_value = JSON.parse(open(fixture_path, &:read))
   result = Maze::Compare.value(expected_value, payload_value)
@@ -30,7 +30,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] Path to the tested element
 # @step_input fixture_path [String] Path to a JSON fixture
-Then('the {word} payload field {string} matches the JSON fixture in {string}') \
+Then('the {request_type}(|a b) payload field {string} matches the JSON fixture in {string}') \
 do |request_type, field_path, fixture_path|
   list = Maze::Server.list_for(request_type)
   payload_value = Maze::Helper.read_key_path(list.current[:body], field_path)
@@ -44,7 +44,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] Path to the tested element
-Then('the {word} payload field {string} is true') do |request_type, field_path|
+Then('the {request_type} payload field {string} is true') do |request_type, field_path|
   list = Maze::Server.list_for(request_type)
   Maze.check.true(Maze::Helper.read_key_path(list.current[:body], field_path))
 end
@@ -53,7 +53,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] Path to the tested element
-Then('the {word} payload field {string} is false') do |request_type, field_path|
+Then('the {request_type} payload field {string} is false') do |request_type, field_path|
   list = Maze::Server.list_for(request_type)
   Maze.check.false(Maze::Helper.read_key_path(list.current[:body], field_path))
 end
@@ -62,7 +62,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] Path to the tested element
-Then('the {word} payload field {string} is null') do |request_type, field_path|
+Then('the {request_type} payload field {string} is null') do |request_type, field_path|
   list = Maze::Server.list_for(request_type)
   Maze.check.nil(Maze::Helper.read_key_path(list.current[:body], field_path))
 end
@@ -71,7 +71,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] Path to the tested element
-Then('the {word} payload field {string} is not null') do |request_type, field_path|
+Then('the {request_type} payload field {string} is not null') do |request_type, field_path|
   list = Maze::Server.list_for(request_type)
   Maze.check.not_nil(Maze::Helper.read_key_path(list.current[:body], field_path))
 end
@@ -81,7 +81,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] Path to the tested element
 # @step_input int_value [Integer] The value to test against
-Then('the {word} payload field {string} equals {int}') do |request_type, field_path, int_value|
+Then('the {request_type} payload field {string} equals {int}') do |request_type, field_path, int_value|
   Maze.check.equal(int_value,
                    Maze::Helper.read_key_path(Maze::Server.list_for(request_type).current[:body], field_path))
 end
@@ -91,7 +91,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] Path to the tested element
 # @step_input float_value [Float] The value to test against
-Then('the {word} payload field {string} equals {float}') do |request_type, field_path, float_value|
+Then('the {request_type} payload field {string} equals {float}') do |request_type, field_path, float_value|
   Maze.check.equal(float_value,
                    Maze::Helper.read_key_path(Maze::Server.list_for(request_type).current[:body], field_path))
 end
@@ -102,7 +102,7 @@ end
 # @step_input field_path [String] Path to the tested element
 # @step_input float_value [Float] The value to test against
 # @step_input places [Int] The number of decimal places to round the actual value to first
-Then('the {word} payload field {string} equals {float} to {int} decimal place(s)') do |request_type, field_path, float_value, places|
+Then('the {request_type} payload field {string} equals {float} to {int} decimal place(s)') do |request_type, field_path, float_value, places|
   body = Maze::Server.list_for(request_type).current[:body]
   rounded_value = Maze::Helper.read_key_path(body, field_path).round places
   Maze.check.equal(float_value,
@@ -114,7 +114,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] The payload element to test
 # @step_input env_var [String] The environment variable to test against
-Then('the {word} payload field {string} equals the environment variable {string}') \
+Then('the {request_type} payload field {string} equals the environment variable {string}') \
 do |request_type, field_path, env_var|
   environment_value = ENV[env_var]
   Maze.check.false(environment_value.nil?, "The environment variable #{env_var} must not be nil")
@@ -129,7 +129,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] The payload element to test
 # @step_input int_value [Integer] The value to compare against
-Then('the {word} payload field {string} is greater than {int}') do |request_type, field_path, int_value|
+Then('the {request_type} payload field {string} is greater than {int}') do |request_type, field_path, int_value|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field_path)
   Maze.check.kind_of Integer, value
@@ -141,7 +141,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] The payload element to test
 # @step_input int_value [Integer] The value to compare against
-Then('the {word} payload field {string} is less than {int}') do |request_type, field_path, int_value|
+Then('the {request_type} payload field {string} is less than {int}') do |request_type, field_path, int_value|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field_path)
   Maze.check.kind_of Integer, value
@@ -154,7 +154,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] The payload element to test
 # @step_input string_value [String] The string to test against
-Then('the {word} payload field {string} equals {string}') do |request_type, field_path, string_value|
+Then('the {request_type} payload field {string} equals {string}') do |request_type, field_path, string_value|
   list = Maze::Server.list_for(request_type)
   Maze.check.equal(string_value, Maze::Helper.read_key_path(list.current[:body], field_path))
 end
@@ -164,7 +164,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] The payload element to test
 # @step_input string_value [String] The string to test against
-Then('the {word} payload field {string} starts with {string}') do |request_type, field_path, string_value|
+Then('the {request_type} payload field {string} starts with {string}') do |request_type, field_path, string_value|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field_path)
   Maze.check.kind_of String, value
@@ -179,7 +179,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field_path [String] The payload element to test
 # @step_input string_value [String] The string to test against
-Then('the {word} payload field {string} ends with {string}') do |request_type, field_path, string_value|
+Then('the {request_type} payload field {string} ends with {string}') do |request_type, field_path, string_value|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field_path)
   Maze.check.kind_of String, value
@@ -194,7 +194,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field [String] The payload element to test
 # @step_input count [Integer] The value expected
-Then('the {word} payload field {string} is an array with {int} elements') do |request_type, field, count|
+Then('the {request_type} payload field {string} is an array with {int} elements') do |request_type, field, count|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field)
   Maze.check.kind_of Array, value
@@ -205,7 +205,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field [String] The payload element to test
-Then('the {word} payload field {string} is a non-empty array') do |request_type, field|
+Then('the {request_type} payload field {string} is a non-empty array') do |request_type, field|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field)
   Maze.check.kind_of Array, value
@@ -217,7 +217,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field [String] The payload element to test
 # @step_input regex [String] The regex to test against
-Then('the {word} payload field {string} matches the regex {string}') do |request_type, field, regex_string|
+Then('the {request_type} payload field {string} matches the regex {string}') do |request_type, field, regex_string|
   regex = Regexp.new(regex_string)
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field)
@@ -228,7 +228,7 @@ end
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input field [String] The payload element to test
-Then('the {word} payload field {string} is a parsable timestamp in seconds') do |request_type, field|
+Then('the {request_type} payload field {string} is a parsable timestamp in seconds') do |request_type, field|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], field)
   begin
@@ -245,7 +245,7 @@ end
 # @step_input request_type [String] The type of request (error, session, build, etc)
 # @step_input key_path [String] The path to the tested array
 # @step_input element_key_path [String] The key for the expected element inside the array
-Then('each element in {word} payload field {string} has {string}') do |request_type, key_path, element_key_path|
+Then('each element in {request_type} payload field {string} has {string}') do |request_type, key_path, element_key_path|
   list = Maze::Server.list_for(request_type)
   value = Maze::Helper.read_key_path(list.current[:body], key_path)
   Maze.check.kind_of Array, value